Move command inside topic test
